Beginning refactoring
