+++ app/src/main/java/com/wikijourney/wikijourney/functions/POI.java
@@ -1,5 +1,7 @@
 package com.wikijourney.wikijourney.functions;
 
+import android.graphics.drawable.Drawable;
+
 /**
  * Created by Thomas on 25/07/2015.
  * This is the class used to store the information we get for each POI using the WikiJourney API.
@@ -14,6 +16,8 @@ public class POI {
     private String type_name;
     private int type_id;
     private int id;
+    private Drawable image;
+    private String description;
 
     public POI(double latitude, double longitude, String name, String sitelink, String type_name, int type_id, int id) {
         this.latitude = latitude;
@@ -79,5 +83,21 @@ public class POI {
 
     public void setId(int id) {
         this.id = id;
+    }
+
+    public Drawable getImage() {
+        return image;
+    }
+
+    public void setImage(Drawable image) {
+        this.image = image;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
     }
 }
