+++ app/src/main/res/layout/fragment_home.xml
@@ -1,7 +1,7 @@
 <FrameLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.wikijourney.wikijourney.fragments.HomeFragment"
+    tools:context="com.wikijourney.wikijourney.views.HomeFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:paddingTop="@dimen/activity_vertical_margin"
