+++ app/src/main/res/layout/fragment_map.xml
@@ -1,7 +1,7 @@
 <FrameLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.wikijourney.wikijourney.fragments.MapFragment"
+    tools:context="com.wikijourney.wikijourney.views.MapFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:paddingTop="@dimen/activity_vertical_margin"
