+++ app/src/main/res/layout/fragment_poi_list.xml
@@ -1,7 +1,7 @@
 <FrameLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.wikijourney.wikijourney.fragments.PoiListFragment"
+    tools:context="com.wikijourney.wikijourney.views.PoiListFragment"
     android:id="@+id/poi_list"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
