+++ app/src/main/res/values/strings.xml
@@ -29,7 +29,7 @@
     <!-- Strings for Options -->
 
     <!-- Strings for About -->
-    <string name="app_version">Alpha 0.0.2</string>
+    <string name="app_version">Alpha 0.0.3</string>
     <string name="about_opensource">WikiJourney is an OpenSource project using data from Wikipedia, Wikidata, WikiVoyage and OpenStreetMaps.</string>
     <string name="about_licenses">Source Code License : Apache 2.0\nContent License : CC-by-SA</string>
     <string name="about_links">Website : http://wikijourney.eu/ \nGithub : https://github.com/WikiJourney/</string>
